backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
httpx==0.27.2
openai==1.51.0
python-dotenv==1.0.1
pydantic==2.9.2
pdfplumber==0.11.4
faiss-cpu==1.13.2
numpy
